requirements4cluster.txt
pip
accelerate==0.24.1  
aiohttp==3.9.0   
aiosignal==1.3.1   
appnope==0.1.3   
asttokens==2.4.1    
attrs==23.1.0   
certifi==2023.11.17 
charset-normalizer==3.3.2 
comm==0.2.0 
datasets==2.15.0 
debugpy==1.8.0
decorator==5.1.1
evaluate==0.4.1
executing==2.0.1
filelock==3.13.1
frozenlist==1.4.0
fsspec==2023.10.0
huggingface-hub==0.19.4
idna==3.4
ipykernel==6.27.0    
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.5.0 
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
nest-asyncio==1.5.8
numpy==1.26.2
packaging==23.2
pandas==2.1.3
parso==0.8.3
peft==0.6.2
pexpect==4.8.0
platformdirs==4.0.0
prompt-toolkit==3.0.41
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.1
pyarrow-hotfix==0.6
Pygments==2.17.2
python-dateutil==2.8.2
pytz==2023.3.post1
pyzmq==25.1.1
regex==2023.10.3
requests==2.31.0
responses==0.18.0
safetensors==0.4.0
scipy==1.11.4
setuptools==68.2.2
six==1.16.0
sklearn==0.0.post11
stack-data==0.6.3
threadpoolctl==3.2.0
tokenizers==0.15.0
traitlets==5.13.0
transformers==4.35.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.12
wheel==0.41.3
xxhash==3.4.1
yarl==1.9.3

#added on 23/12/2023

tabulate==0.8.0
hydra-core==1.3.2
gpytorch==1.11  
omegaconf==2.3.0